requirements.txt
python-dotenv==0.19.2
aiogram==2.20
aiohttp==3.8.1